requirements.txt
streamlit>=1.36.0
openai>=1.40.0
python-dotenv>=1.0.1
fdb>=2.0.2
